# tt-media-server/cpp_server/resources/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TT Media Server API (C++ Drogon)",
    "description": "High-performance C++ implementation of the TT Media Server using Drogon framework. Provides OpenAI-compatible chat completions API for benchmarking server overhead.",
    "version": "1.0.0",
    "contact": {
      "name": "Tenstorrent",
      "url": "https://tenstorrent.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Local server"
    }
  ],
...
  "paths": {
    "/v1/chat/completions": {
      "post": {
        "tags": ["Chat Completions"],
        "summary": "Create chat completion",
        "description": "Creates a chat completion for the provided messages. OpenAI-compatible endpoint. Messages are converted to a prompt internally; responses use object \"chat.completion\" and choices[].message with role and content.",
        "operationId": "createChatCompletion",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatCompletionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful chat completion (JSON) or streaming (text/event-stream)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChatCompletionResponse" }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "SSE stream; object \"chat.completion.chunk\", choices[].delta"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g. missing or empty messages)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "503": {
            "description": "Model not ready",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["messages"],
        "properties": {
          "model": {
            "type": "string",
            "description": "Model identifier",
            "example": "test-model"
          },
          "messages": {
            "type": "array",
            "description": "Conversation messages (required, non-empty)",
            "items": {
              "type": "object",
              "required": ["role", "content"],
              "properties": {
                "role": {
                  "type": "string",
                  "description": "Message role: system, user, or assistant",
                  "default": "user"
                },
                "content": {
                  "type": "string",
                  "description": "Message content"
                }
              }
            }
          },
          "max_tokens": {
            "type": "integer",
            "default": 16,
            "minimum": 1,
            "description": "Maximum number of tokens to generate"
          },
          "stream": {
            "type": "boolean",
            "default": false,
            "description": "Whether to stream the response as SSE"
          },
          "stream_options": {
            "$ref": "#/components/schemas/StreamOptions"
          },
          "temperature": {
            "type": "number",
            "minimum": 0,
            "maximum": 2,
            "description": "Sampling temperature"
          },
          "top_p": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "Nucleus sampling probability"
          },
          "stop": {
            "oneOf": [
              { "type": "string" },
              { "type": "array", "items": { "type": "string" } }
            ],
            "description": "Stop sequence(s)"
          },
          "presence_penalty": { "type": "number" },
          "frequency_penalty": { "type": "number" },
          "seed": { "type": "integer" },
          "user": { "type": "string" },
          "response_format": {
            "$ref": "#/components/schemas/ResponseFormat"
          }
        }
      },
      "ResponseFormat": {
        "type": "object",
        "required": ["type"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["text", "json_object", "json_schema"],
            "description": "The format the model must output. 'text' for free-form, 'json_object' for any valid JSON, 'json_schema' for JSON matching a specific schema."
          },
          "json_schema": {
            "type": "object",
            "description": "Required when type is 'json_schema'. Specifies the JSON schema the output must conform to.",
            "properties": {
              "name": {
                "type": "string",
                "description": "A name for the schema (for identification)"
              },
              "strict": {
                "type": "boolean",
                "default": false,
                "description": "Whether to enforce strict schema adherence"
              },
              "schema": {
                "type": "object",
                "description": "The JSON Schema object that the output must match"
              }
            },
            "required": ["schema"]
          }
        }
      },
      "ChatCompletionResponse": {
        "type": "object",
        "required": ["id", "object", "created", "choices"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique chat completion identifier",
            "example": "chatcmpl-abc123def456"
          },
          "object": {
            "type": "string",
            "enum": ["chat.completion"],
            "description": "Object type"
          },
          "created": {
            "type": "integer",
            "description": "Unix timestamp of creation"
          },
          "model": {
            "type": "string",
            "description": "Model used for completion"
          },
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": { "type": "integer" },
                "message": {
                  "type": "object",
                  "properties": {
                    "role": { "type": "string" },
                    "content": { "type": "string" }
                  }
                },
                "finish_reason": { "type": "string" }
              }
            }
          },
          "usage": {
            "$ref": "#/components/schemas/CompletionUsage"
          }
        }
      },
      "StreamOptions": {
        "type": "object",
        "properties": {
          "include_usage": {
            "type": "boolean",
            "default": true,
            "description": "Include usage statistics in response"
          },
          "continuous_usage_stats": {
            "type": "boolean",
            "default": false,
            "description": "Include usage stats in each streamed chunk"
          }
        }
      },
      "CompletionUsage": {
        "type": "object",
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "description": "Number of tokens in the prompt"
          },
          "completion_tokens": {
            "type": "integer",
            "description": "Number of tokens in the completion"
          },
          "total_tokens": {
            "type": "integer",
            "description": "Total tokens used"
          },
          "ttft_ms": {
            "type": "number",
            "description": "Time to first token in milliseconds",
            "nullable": true
          },
          "tps": {
            "type": "number",
            "description": "Tokens per second (excluding first token)",
            "nullable": true
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "type": { "type": "string" },
              "code": { "type": "string" }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Bearer token authentication using OPENAI_API_KEY environment variable. If not set, defaults to 'your-secret-key'."
      }
    }
  }
}